pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fin-parser"
version = "0.1.0"
description = "AI-powered financial filing parser using EDGAR + Claude"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Phase 1 — Ingestion + extraction
    "requests>=2.31",
    "pdfplumber>=0.11",
    "anthropic>=0.25",
    "python-dotenv>=1.0",

    # Phase 2 — Valuation + analysis
    "numpy>=1.26",
    "numpy-financial>=1.0",
    "pandas>=2.2",

    # Phase 3 — Dashboard + embeddings
    "streamlit>=1.35",
    "plotly>=5.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipython>=8.0",
]

[project.scripts]
fin-parser = "fin_parser.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=fin_parser --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["."]
include = ["fin_parser*"]
